requirements.txt
# https://docs.llamaindex.ai/en/stable/getting_started/installation/#custom-installation-from-pip
llama-index-core
llama-index-embeddings-ollama
llama-index-llms-ollama
ollama
chromadb
llama-index-vector-stores-chroma
